+++ requirements.in
@@ -7,7 +7,6 @@ pytest-docstyle==1.5.0
 pytest-flake8==1.0.6
 pytest-mypy==0.4.2
 pytest-xdist==1.31.0
-pytest-dependency==0.5.1
 pytest==5.4.1
 PyYAML==5.3
 smart_open==1.10.0
@@ -18,5 +17,3 @@ cattrs==1.0.0
 attrs==19.3.0
 typing==3.7.4.1
 click==7.1.2
-zipp==3.1.0
-importlib-metadata==1.7.0
